support float label when convert imageset

+++ include/caffe/util/io.hpp
@@ -80,13 +80,13 @@ inline void WriteProtoToBinaryFile(
   WriteProtoToBinaryFile(proto, filename.c_str());
 }
 
-bool ReadFileToDatum(const string& filename, const int label, Datum* datum);
+bool ReadFileToDatum(const string& filename, const float label, Datum* datum);
 
 inline bool ReadFileToDatum(const string& filename, Datum* datum) {
   return ReadFileToDatum(filename, -1, datum);
 }
 
-bool ReadImageToDatum(const string& filename, const int label,
+bool ReadImageToDatum(const string& filename, const float label,
     const int height, const int width, const bool is_color,
     const std::string & encoding, Datum* datum);
